move tests around to avoid big if-s and a dummy test to avoid having a test file without TestCase-s when neither LinguaPlone nor p.a.multilingual are installed (for instance base Plone 4.3.x)
